Clean up for command classes. Use java generics in Cmd class
