+++ src/main/java/duke/command/dishesCommand/AddDishCommand.java
@@ -1,29 +1,54 @@
-package duke.dishesCommand;
+package duke.command.dishesCommand;
 
-import duke.command.AddCommand;
-import duke.exception.DukeException;
 import duke.Dishes.DishList;
+import duke.Dishes.Dishes;
+import duke.command.Cmd;
+import duke.exception.DukeException;
 import duke.storage.Storage;
-import duke.task.TaskList;
 import duke.ui.Ui;
 
-public class AddIngredient extends RecipeCommand {
+public class AddDishCommand extends Cmd<DishList> {
 
-    private String ingredient;
+    private Dishes dish;
+    private int amount;
     private int Nb;
 
-    public AddIngredient(String ingredient, int Nb) {
-        this.ingredient = ingredient;
-        this.Nb = Nb;
+    public AddDishCommand(Dishes dish, int amount) {
+        this.dish = dish;
+        this.amount = amount;
     }
 
     @Override
-    public void execute(DishList dish1, TaskList taskList, Ui ui, Storage storage) throws DukeException {
+    public void execute(DishList dish1, Ui ui, Storage storage) throws DukeException {
+        boolean flag = true;
         try {
-            dish1.getDish(Nb - 1).addIngredients(ingredient);
-            System.out.println("\t added ingredient: " + ingredient + "\n\t to dish: " + dish1.getDish(Nb - 1).getDishname());
+            if(dish1.getSize() == 0) {
+                dish1.addDish(dish);
+                dish1.getDish(0).setNumberOfOrders(amount);
+                ui.showAddedDishes(dish.getDishname(), amount);
+            }
+            else {
+                for( int i = 0; i < dish1.getSize(); i++) {
+                    if(dish1.getDish(i).getDishname().equals(dish.getDishname())){
+                        Nb = i;
+                        flag = false; //dish already exist in list
+                        break;
+                    }
+                }
+                if(flag) {
+                    dish1.addDish(dish); // add dish into list found in dishes class
+                    dish1.getDish(dish1.getSize() - 1).setNumberOfOrders(amount);
+                    ui.showAddedDishes(dish.getDishname(), dish1.getDish(dish1.getSize() - 1).getTotalNumberOfOrders());
+                }
+                else {
+                    dish1.getDish(Nb).setNumberOfOrders(amount);
+                    System.out.println("\t your updated orders:\n\t "
+                            + dish.getDishname() + "\t amount: " +
+                            String.valueOf(dish1.getDish(Nb).getTotalNumberOfOrders()));
+                }
+            }
         } catch (Exception e) {
-            throw new DukeException("cannot add ingredient");
+            throw new DukeException("unable to add dish");
         }
     }
 }
