--- src/main/java/duke/dishesCommand/DeleteDishCommand.java
@@ -1,34 +1,0 @@
-package duke.dishesCommand;
-
-import duke.exception.DukeException;
-import duke.Dishes.DishList;
-import duke.storage.Storage;
-import duke.task.Task;
-import duke.task.TaskList;
-import duke.ui.Ui;
-
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.function.DoubleUnaryOperator;
-
-public class DeleteDishCommand extends RecipeCommand {
-
-    private int Nb;
-
-    public DeleteDishCommand(int dishNb) {
-        this.Nb = dishNb;
-    }
-
-    @Override
-    public void execute(DishList dish1, TaskList taskList, Ui ui, Storage storage) throws DukeException {
-        try {
-            ui.showDeletedDIsh(dish1.getDish(Nb - 1).getDishname());
-            dish1.deleteDish(Nb - 1);
-        } catch (Exception e) {
-            throw new DukeException("dish does not exist");
-        }
-    }
-}
