+++ src/main/java/duke/parser/Parser.java
@@ -5,8 +5,8 @@ import duke.Duke;
 import duke.command.ingredientCommand.*;
 import duke.exception.DukeException;
 import duke.command.orderCommand.*;
-import duke.command.dishesCommand.*;
-import duke.command.dishesCommand.InitCommand;
+import duke.dishesCommand.*;
+import duke.dishesCommand.InitCommand;
 import duke.Dishes.Dishes;
 import duke.task.Deadline;
 import duke.task.DoWithinPeriodTasks;
